chore: temporarily disable XS test
We learned (in #1244) that the XS RegExp engine rejects several of the regexp
literals that appear in the SES shim (via babel-standalone). The XS test
works by compiling some other unit tests (eventual-send and marshal) into a C
executable, then running it. #1244 added `install-ses` to most packages,
including those two, which means XS is now being exposed to SES-shim code
that it wasn't seeing before, triggering the rejection bug.

When XS is updated to tolerate these cases, we'll turn this test back on.

closes #1251