feat(import-bundle)!: add inescapable transforms/globalLexicals options
This adds a compartment-wrapper to attach inescapable options (transforms
and globalLexicals) to the Compartment constructor. Not only will they be
applied to code evaluated in the compartment, but also to code in any
descendant Compartments created by that code.

I was originally planning to publish the wrapper as a separate package in the
SES-shim monorepo (https://github.com/Agoric/SES-shim/issues/377), but it
seems useful to let it bake in `@agoric/import-bundle` (next to its first
customer) for a while instead.

Note that the wrapper assumes that all compartments have equivalent
`Compartment` constructors, which will cease to be the case when SES changes
soon (https://github.com/Agoric/SES-shim/issues/366). This commit adds a
peerDependency upon the version of SES which is known to work: when we
upgrade `@agoric/install-ses` to use the new SES, this should trigger a
warning, and we should heed it. We'll need to rewrite the wrapper at that
point.

refs #1296