chore(swingset): internal consistency on buildRootObject() signature
We were calling dynamic vats with `buildRootObject(vatPowers)` as documented,
but static vats were still getting the old `(E, D, vatPowers)` layout.

This fixes swingset's internals to be consistent, and also moves the built-in
vats from setup() to buildRootObject() whereever possible. This is
facilitated by the addition of `D()` to `vatPowers`.

refs #1233
refs #1111
closes #33