chore(swingset): refactor VatManager interaction
Redefine VatManager according to the new architecture in docs/vat-worker.md .
The translation duties in the old vatManager.js are split out into
"translators", which are built separately for each vat, but which don't care
how the VatWorker is implemented.

Devices are now defined as files with a named `buildRootDeviceNode` function,
just like how vats export a `buildRootObject` function. However device files
no longer support `setup` at all.

This should pave the way for new kinds of VatWorkers, such as XS in a child
process.

I removed some leftover code in `kernel.js` that guarded against cross-Realm
function calls (stringifying arguments, copy arrays into new arrays) because
new-SES means we have only one shared Realm everywhere.

The code in kernel.js that builds the metering tools was factored out to a
separate file.

`vatManager.js` was broken up into a number of separate files, with
`src/kernel/vatManager/vatManager.js` being the entry point. The
inbound/outbound translation code was moved out to `vatTranslator.js`,
because it does not depend upon the particular kind of VatWorker we're using.

The portions of `kernel.js` which handled syscalls on behalf of vats was
moved out to `kernelSyscall.js`, and made to work in terms of `KernelSyscall`
objects (which come out of the translator, which is fed with a `VatSyscall`
object, which comes from the VatManager). Statistics incrementing is the
responsibility of `kernelSyscall.js`.

The `panic()` function was made to take the Error object it will be throwing,
to preserve the stack trace when an exception kills the kernel (as opposed to
an assertion failure).

refs #1321