chore(swingset): refactor vatManagerFactory()
Refactor the creation of VatManagers to use a single
function (vatManagerFactory), which takes "managerOptions", which include
everything needed to configure the new manager.

The handling of options was tightened up, with precondition checks on the
options-bag contents.

One new managerOption is `managerType`, which specifies where we want the new
vat to run. Only `local` is supported right now (our usual Compartment
sharing the main thread approach), but #1299 / #1127 will introduce new types
that use separate threads, or subprocesses.

The controller/kernel acquired a new `shutdown()` method, to use in unit
tests. This is unused now, but once we add workers that spawn threads or
subprocesses, we'll need to call+await this at the end of any unit test that
creates thread/subprocess-based workers, otherwise the worker thread would
keep the process from ever exiting.