feat(swingset-vat): add xs-worker managerType
If any vats are configured to use it, but xs-vat-worker does not have
an executable to offer, swingset will throw an error. (IOU a test?)

 - Swingset package depends on xs-vat-worker
   - ISSUE: cycle. factor out liveslots?
 - add startXsWorker kernel endowment,
   which is undefined unless binary is built
 - parameterize execPath, args in startSubprocessWorker()
   (mantain previous defaults)
 - xs-vat-worker: adopt array style protocol from SwingSet (WIP)
   - based on subprocessSupervisor.js
     copied SwingSet/src/kernel/vatManager/subprocessSupervisor.js
     wholesale; applied ocap discipline to module-level mutable state;
     adapted IO interface. Compare:

Initially it would hang after the first 2 syscalls of the first
deliver:

 - [ 'subscribe', 'p-60' ]
 - [ 'subscribe', 'p-61' ]

Then I fixed handle() to return a promise in all cases to be sure to
finish delivery before next blocking read.