chore: remove/disable a lot of kernel startup debug messages
AVA runs tests in parallel, which makes the kernel startup messages pretty
messy. I don't think they're adding a huge amount of value, so I disabled
most of them. (I left in the "is global metering enabled" for now, the
problem it reveals might be hard to detect otherwise).

I added a `verbose:` option to switch them back on if we really need them.