feat: allow pre-built kernelBundles for faster unit tests
Using `rollup` to turn a tree of vat-defining source files into a bundle
object is somewhat slow. It seems to depend on the amount of code being
bundled: the worst case is the kernel sources, which take 1.5s . The kernel
needs a number of other vats which are automatically bundled at runtime. The
total process takes about 2.7s on my machine.

We have unit tests that run many swingsets in a single test file. The most
extreme case is test-message-patterns.js, which calls `buildVatController` 60
times in a row. It takes a really long time to run, 3-6 minutes, and the
majority is the bundling step. Parallelizing this doesn't help (JS is a
single-threaded language, it's not using non-blocking filesystem operations,
I don't know whether it's CPU or IO bound but either would be a problem).

This patch exports a new `buildKernelBundles()` function to let these
multi-swingset tests build the bundles once per test file. It also adds
`kernelBundles` to the `buildVatController(.., .., runtimeOptions)` options
bag to pass these bundles back in (with sensible defaults). The controller
gets all the built-in bundles from this collection, rather than running
`bundleSource` each time.

It also rearranges the console setup slightly, to make the `console` object
available earlier within `buildVatController`. I needed this to debug
problems with unpacking the other `runtimeOptions`. Also, now that we're
building our own `console` object, we no longer need to explicitly harden the
native console's (unique) prototype object.

refs #1643