fix: don't modify the original 'config' object
To support `buildVatController(.., argv)`, the code modified the `config`
object (replacing `.parameters` on the bootstrap vat with the argv contents).
This mutation of the inbound `config` argument is a bit surprising, and
interferes with multi-swingset unit tests that might want to share the config
object among multiple `buildVatController` calls, or if the `config` object
is hardened.

This patch makes the controller build a new `config` object out of the pieces
of the original, with the new parameters overlaid on top.

Provides most of the fix to #1490 , but doesn't close it, since we still have
to think about `.devices`.