test: add 'test-nyc' command to gather/report coverage data
This adds 'nyc' to our devDependencies, along with a patch (installed with
patch-package) to make it compatible with SES. We configure it with
.nycrc.yaml, and add `yarn test-nyc` commands to all packages that have unit
tests which might work with NYC.

'nyc' (and 'istanbul', the coverage-injecting library) does not know how to
work with import-bundle, so tests that use swingset or other forms of
import-bundle (e.g. spawner) won't report the bundled code as being executed,
making the coverage numbers look pretty bad.

Some tests report errors when run as `yarn test-nyc`: bundle-source,
import-bundle, swingset-runner, and spawner. This seems to involve the
injected instrumentation not running against the global object that it
expects. For this reason, I haven't added any top-level "run all tests with
coverage" command yet. Developers should run `yarn test-nyc` (or `yarn
test-nyc test/test-FILE.js`) in single packages, to improve the utility of
their tests.

At some point in the future, we'll add a global command, and begin
collecting/reporting coverage metrics.

refs #1664