fix(xsnap): Iron out resolution types
The xsnap controller passes a "baton" from the completion of one interaction with the xsnap subprocess to the next, regardless of whether the interaction completed sucessfully.  The xnsap controller can also kill the subprocess, and the child can terminate unexpectedly, so every interaction is a race to completion or the child process exiting.

In a prior solution to the baton passing, the result of an interaction could be a rejection. To allow the next interaction to commence, the baton would commence with the conclusion of the prior interaction, and if that produced an exception, the rejection would be caught and transformed into a resolution to an error.

This is weird.

This change makes both the child process exit and the baton promise the same type: Promise<void> and termination is always communicated as a rejection.

This change also adds a `terminate` method to allow a distinction between graceful close and forced termination.