fix(swingset): loadVat.js: properly wait for static vats to be ready
As we discovered while fixing #2213, `recreateStaticVat()` was not returning
the Promise needed by the kernel's `start()` function to wait for all vats to
be ready for messages, before it proceeds with the bootstrap crank. We got
lucky for managerType=`local`, because the vats were finished building before
any messages were delivered. But for a type of `nodeWorker` or
`node-subprocess`, it takes longer for `build()` to complete. As a result,
the vatManager was not yet added to `ephemeral.vats` by the time the kernel
attempted to deliver a message to the latecomer, which made the kernel think
that vat had been terminated.