test: hush logged Errors by using NonErrors (#2392)
The marshal module emits a warning (with stack trace) to the console each
time it serializes an Error, which makes it look like tests are failing. We
have tests which test 'raise' and Promise rejection to make sure they are
signalled correctly. We previously used 'raise Error()' for this, but that
provokes the scary-looking warning. Since we aren't trying to exercise
*Error* serialization here, merely Promise rejection, we can use a non-Error
instead, which avoids the warning.

We use a pass-by-copy object, so the receiving side can log a quoted
(JSON.stringify) version, and not wind up with Presence object-ids in the
log. This does require changes to how the tests log "errors", and to the
strings we compare those logs against.

refs #2368, but we should hold off closing it until we find all the other
swingset tests that use this pattern and change them too