chore(swingset): remove decref support
This was a false start on the GC task. Previously we thought vat-side
finalizers would call a vat-specific `decref` function at spontaneous times,
and the kernel would accumulate the dropped references for later processing.
Now we have liveslots collect the finalizer results and perform a
non-transcript-enforced syscall.dropImports at the end of crank.

This removes `testTrackDecref` from `kernelOptions`.