test(ava-xs): integrate with ses-ava, install-ses-debug
install-ses-debug installs SES and calls lockdown() with debugging
options. In ava-xs, we avoid installing SES again (since it's already
installed) but we have to set the lockdown() options to match so
that things like error messsages match test expectations.

So we make another copy of the SES bootstrap bundle that calls
lockdown with debugging options.

We also supply a stub `@agoric/ses-ava`.